Refactored library.app module; deleted some unused classes.
