+++ headSimulatorTwo/src/main/java/app/TheSubscriber.java
@@ -1,4 +1,4 @@
-package app.library;
+package app;
 
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
